_site/requirements.txt
python-a2a
PyYAML
feedparser
requests
jenkins 
groq
python-dotenv
aiohttp
